🔥 PoC 코드 삭제.
