--- state/src/commonMain/kotlin/kr/lul/stringnotebook/state/organism/NotebookPreferences.kt
@@ -1,11 +1,0 @@
-package kr.lul.stringnotebook.state.organism
-
-import androidx.compose.runtime.Stable
-
-/**
- * 노트북 설정.
- */
-@Stable
-interface NotebookPreferences {
-    object Default : NotebookPreferences
-}
