+++ state/src/commonMain/kotlin/kr/lul/stringnotebook/state/page/MainPageState.kt
@@ -1,10 +1,6 @@
 package kr.lul.stringnotebook.state.page
 
 import androidx.compose.runtime.Immutable
-import kr.lul.stringnotebook.domain.foundation.EventProcessor
-import kr.lul.stringnotebook.state.organism.Context
-import kr.lul.stringnotebook.state.organism.NotebookFocusedContext
-import kr.lul.stringnotebook.state.organism.NotebookState
 import kotlin.uuid.ExperimentalUuidApi
 
 /**
@@ -14,8 +10,5 @@ import kotlin.uuid.ExperimentalUuidApi
  */
 @ExperimentalUuidApi
 @Immutable
-data class MainPageState(
-    val notebook: NotebookState = NotebookState(),
-    val context: Context = NotebookFocusedContext(),
-    val processor: EventProcessor = EventProcessor.NoOp
+class MainPageState(
 )
