+++ ui/preview/src/androidMain/kotlin/kr/lul/stringnotebook/preview/ui/foundation/ComponentPreviewContainer.kt
@@ -1,23 +1,20 @@
-package kr.lul.stringnotebook.preview.ui.atom
+package kr.lul.stringnotebook.preview.ui.foundation
 
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import kr.lul.stringnotebook.ui.atom.StringNotebookTheme
 
 @Composable
-fun PagePreviewContainer(
+fun ComponentPreviewContainer(
     modifier: Modifier = Modifier,
+    contentAlignment: Alignment = Alignment.Center,
     content: @Composable () -> Unit
 ) {
     StringNotebookTheme {
-        Scaffold {
-            Box(modifier.padding(it), Alignment.Center) {
-                content()
-            }
+        Box(modifier, contentAlignment) {
+            content()
         }
     }
 }
