--- ui/src/commonMain/kotlin/kr/lul/stringnotebook/ui/atom/PointerInputChangeExtension.kt
@@ -1,6 +1,0 @@
-package kr.lul.stringnotebook.ui.atom
-
-import androidx.compose.ui.input.pointer.PointerInputChange
-
-val PointerInputChange.summary: String
-    get() = "(id=$id, type=$type, position=$position, pressed=$pressed)"
